backend/requirements.txt
aiohttp==3.10.5
numpy==1.26.4
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.5
linkedin-api==2.0.0
composio-langchain==0.5.23